Use EuiButtonContent in EuiButtonEmpty
